requirements.txt
python-telegram-bot==20.7
aiofiles==23.2.1
aiosqlite==0.19.0
python-dotenv==1.0.0
Pillow==10.1.0
qrcode==7.4.2